Moved Boundary Event Common Behaviour method to support folder
