+++ tests/e2e/specs/BoundaryMessageEvent.spec.js
@@ -5,7 +5,7 @@ import {
   setBoundaryEvent,
 } from '../support/utils';
 import { nodeTypes } from '../support/constants';
-import { CommonBoundaryEventBehaviour } from './BoundaryEventCommonBehaviour.spec';
+import { CommonBoundaryEventBehaviour } from '../support/BoundaryEventCommonBehaviour';
 
 describe('Boundary Message Event', () => {
   const taskPosition = { x: 200, y: 200 };
