+++ tests/e2e/specs/BoundarySignalEvent.spec.js
@@ -7,7 +7,7 @@ import {
   waitToRenderAllShapes,
 } from '../support/utils';
 import { nodeTypes } from '../support/constants';
-import { CommonBoundaryEventBehaviour } from './BoundaryEventCommonBehaviour.spec';
+import { CommonBoundaryEventBehaviour } from '../support/BoundaryEventCommonBehaviour';
 
 describe.skip('Boundary Signal Event', () => {
   const taskPosition = { x: 200, y: 200 };
